Semi-refactor the sidebar
